Cleaned up. Minor things.
