--- temp/output.yaml
@@ -1,37 +1,0 @@
-name: mo-582f7d80-a920-4542-a962-978d2d9f443e
-num_nodes: 1
-resources:
-  infra: aws/us-east-1
-  use_spot: true
-  disk_size: 500GB
-  instance_type: g6e.xlarge
-  ports: 8001
-file_mounts:
-  /data:
-    mode: MOUNT
-    source: s3://tandemn-orca
-setup: 'export PYTHONHASHSEED=0
-
-  set -euxo pipefail
-
-  python3 -m pip install -U pip
-
-  python3 -m pip install -U uv
-
-
-  uv venv --python 3.12 --seed
-
-  source .venv/bin/activate
-
-
-  # Install vLLM
-
-  uv pip install "datasets" "requests"
-
-  uv pip install "vllm==0.10.0" # old vllm
-
-  '
-run: "set -euxo pipefail\nsource .venv/bin/activate\n\n~/sky_templates/ray/start_cluster\n\
-  \nif [ \"$SKYPILOT_NODE_RANK\" == \"0\" ]; then\n    vllm run-batch -i /data/orange/test.jsonl\
-  \ -o /data/output_out.txt --model Qwen/Qwen3-0.6B \\\n    -tp 1 -pp 1 --enable-metrics\
-  \ --url 0.0.0.0 --port 8001 \\\n    \nfi"
